+++ tx-backend/src/main/resources/db/migration/V18__remove_fields_from_notificationmessage.sql
@@ -1,0 +1,5 @@
+ALTER TABLE investigation_notification DROP COLUMN is_initial;
+ALTER TABLE alert_notification DROP COLUMN is_initial;
+
+ALTER TABLE investigation_notification DROP COLUMN edc_url;
+ALTER TABLE alert_notification DROP COLUMN edc_url;